修改字符串转color的bug；删除UIImage+addtion大部分重复的代码；新增证件号判断功能的封装

+++ Categories/UIColor+Expanded.m
@@ -508,7 +508,7 @@ static NSLock *crayolaNameCacheLock;
 	NSScanner *scanner = [NSScanner scannerWithString:stringToConvert];
 	if (![scanner scanString:@"{" intoString:NULL]) return nil;
 	const NSUInteger kMaxComponents = 4;
-	CGFloat c[kMaxComponents];
+	float c[kMaxComponents];
 	NSUInteger i = 0;
 	if (![scanner scanFloat:&c[i++]]) return nil;
 	while (1) {
@@ -543,7 +543,7 @@ static NSLock *crayolaNameCacheLock;
     NSScanner *scanner = [NSScanner scannerWithString:stringToConvert];
 	if (![scanner scanString:@"{" intoString:NULL]) return nil;
 	const NSUInteger kMaxComponents = 4;
-	CGFloat c[kMaxComponents];
+    float c[kMaxComponents];
 	NSUInteger i = 0;
 	if (![scanner scanFloat:&c[i++]]) return nil;
 	while (1) {
